requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.4.2
python-dateutil==2.8.2
geopy==2.4.0
scipy==1.11.3
matplotlib==3.7.2
networkx==3.1
flask-socketio==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
scikit-learn==0.24.2
seaborn==0.11.2
tqdm==4.62.3
pytest==8.0.2
black==24.2.0
isort==5.13.2
flake8==7.0.0
eventlet==0.33.3